#include<stdio.h> #include<math.h> int main() {     int p,r,t;     scanf("%d%d%d",&p,&r,&t); float ci; ci=p*pow(1+(r/100.0),t)-p; printf("%0.2f",ci);
}

+++ Compound_Interest.c
@@ -1,14 +1,8 @@
-//COMPOUND INTEREST (FIRST LOGIC BEGINNER – BASIC PROGRAMME)
 #include<stdio.h>
 #include<math.h>
 int main()
 {
-    float p,r,t;
-    scanf("%f%f%f",&p,&r,&t);
-    float ci;
-    ci=p*pow(1+(r/100),t)-p;
-    printf("%0.2f",ci);
+    int p,r,t;
+    scanf("%d%d%d",&p,&r,&t); float ci; ci=p*pow(1+(r/100.0),t)-p; printf("%0.2f",ci);
+    
 }
-
-
-
